requirements.txt
PyQt6>=6.4.0
PyFingerprint>=1.5
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.3.1
pyserial>=3.5
scikit-learn>=0.24.2
seaborn>=0.11.2
matplotlib>=3.4.3
numpy>=1.21.2
tqdm>=4.62.3 